Check Presence of /index for reexports
When processing a reexport (AllExport), it is not sufficent to check for
a sourceLib with the path available in AllExport.from.
To ensure we do not miss any declarations, it is necessary to check for
resources with a '/index' postfix too.

This should fix (or at least be a step towards fixing):
* #226
* #175
* #208